♻️ Refactor: 상수 분리, 전체에 해당하는 폴더 name, id 상수화
