+++ .eslintrc.json
@@ -23,6 +23,7 @@
     }
   },
   "rules": {
+    "@typescript-eslint/no-unused-vars": "off",
     "prettier/prettier": "error",
     "react/jsx-uses-react": "off",
     "react/react-in-jsx-scope": "off",
