+++ components/FolderManagementMenu/index.tsx
@@ -2,6 +2,7 @@ import { folderManagementButtons } from './folderManagementButtons';
 import { FolderManagementButton } from './FolderManagementButton';
 import styles from './styles.module.css';
 import { SelectedFolder } from '@/types/Common';
+import { ALL_CONTENTS_FOLDER } from '@/constants/constants';
 
 interface Props {
   selectedFolder: SelectedFolder;
@@ -13,7 +14,7 @@ export const FolderManagementMenu = ({ selectedFolder }: Props) => {
       <div className={styles['selected-folder__menu']}>
         {selectedFolder.name}
       </div>
-      {selectedFolder.id !== 'all' && (
+      {selectedFolder.id !== ALL_CONTENTS_FOLDER.ID && (
         <div className={styles['folder-management-buttons']}>
           {folderManagementButtons.map(button => {
             return (
