+++ constants/time.ts
@@ -1,0 +1,12 @@
+export const TIMES = {
+  ONE_MINUTE_AGO: '1 minute ago',
+  MINUTES_AGO: ' minutes ago',
+  ONE_HOUR_AGO: '1 hour ago',
+  HOURS_AGO: ' hours ago',
+  ONE_DAY_AGO: '1 day ago',
+  DAYS_AGO: ' days ago',
+  ONE_MONTH_AGO: '1 month ago',
+  MONTHS_AGO: ' months ago',
+  ONE_YEAR_AGO: '1 year ago',
+  YEARS_AGO: ' years ago',
+} as const;
